Updated helper construction
